+++ Source/SwiftLintFramework/Protocols/Rule.swift
@@ -10,11 +10,11 @@ import SourceKittenFramework
 
 public protocol Rule {
     init() // Rules need to be able to be initialized with default values
-    init(config: AnyObject) throws
+    init(configuration: AnyObject) throws
     static var description: RuleDescription { get }
     func validateFile(file: File) -> [StyleViolation]
     func isEqualTo(rule: Rule) -> Bool
-    var configDescription: String { get }
+    var configurationDescription: String { get }
 }
 
 extension Rule {
@@ -25,32 +25,32 @@ extension Rule {
 
 public protocol OptInRule: Rule {}
 
-public protocol ConfigProviderRule: Rule {
-    typealias ConfigType: RuleConfig
-    var config: ConfigType { get set }
+public protocol ConfigurationProviderRule: Rule {
+    typealias ConfigurationType: RuleConfiguration
+    var configuration: ConfigurationType { get set }
 }
 
 public protocol CorrectableRule: Rule {
     func correctFile(file: File) -> [Correction]
 }
 
-// MARK: - ConfigProviderRule conformance to Configurable
+// MARK: - ConfigurationProviderRule conformance to Configurable
 
-public extension ConfigProviderRule {
-    public init(config: AnyObject) throws {
+public extension ConfigurationProviderRule {
+    public init(configuration: AnyObject) throws {
         self.init()
-        try self.config.setConfig(config)
+        try self.configuration.applyConfiguration(configuration)
     }
 
     public func isEqualTo(rule: Rule) -> Bool {
         if let rule = rule as? Self {
-            return config.isEqualTo(rule.config)
+            return configuration.isEqualTo(rule.configuration)
         }
         return false
     }
 
-    public var configDescription: String {
-        return config.consoleDescription
+    public var configurationDescription: String {
+        return configuration.consoleDescription
     }
 }