+++ Source/SwiftLintFramework/Rules/ControlStatementRule.swift
@@ -8,9 +8,9 @@
 
 import SourceKittenFramework
 
-public struct ControlStatementRule: ConfigProviderRule {
+public struct ControlStatementRule: ConfigurationProviderRule {
 
-    public var config = SeverityConfig(.Warning)
+    public var configuration = SeverityConfiguration(.Warning)
 
     public init() {}
 
@@ -67,7 +67,7 @@ public struct ControlStatementRule: ConfigProviderRule {
                     return nil
                 }
                 return StyleViolation(ruleDescription: self.dynamicType.description,
-                    severity: self.config.severity,
+                    severity: self.configuration.severity,
                     location: Location(file: file, characterOffset: match.location))
             }
         }