+++ Source/SwiftLintFramework/Rules/FileLengthRule.swift
@@ -8,8 +8,8 @@
 
 import SourceKittenFramework
 
-public struct FileLengthRule: ConfigProviderRule {
-    public var config = SeverityLevelsConfig(warning: 400, error: 1000)
+public struct FileLengthRule: ConfigurationProviderRule {
+    public var configuration = SeverityLevelsConfiguration(warning: 400, error: 1000)
 
     public init() {}
 
@@ -27,11 +27,11 @@ public struct FileLengthRule: ConfigProviderRule {
 
     public func validateFile(file: File) -> [StyleViolation] {
         let lineCount = file.lines.count
-        for parameter in config.params where lineCount > parameter.value {
+        for parameter in configuration.params where lineCount > parameter.value {
             return [StyleViolation(ruleDescription: self.dynamicType.description,
                 severity: parameter.severity,
                 location: Location(file: file.path, line: lineCount),
-                reason: "File should contain \(config.warning) lines or less: " +
+                reason: "File should contain \(configuration.warning) lines or less: " +
                         "currently contains \(lineCount)")]
         }
         return []