+++ Source/SwiftLintFramework/Rules/ForceTryRule.swift
@@ -8,9 +8,9 @@
 
 import SourceKittenFramework
 
-public struct ForceTryRule: ConfigProviderRule {
+public struct ForceTryRule: ConfigurationProviderRule {
 
-    public var config = SeverityConfig(.Error)
+    public var configuration = SeverityConfiguration(.Error)
 
     public init() {}
 
@@ -29,7 +29,7 @@ public struct ForceTryRule: ConfigProviderRule {
     public func validateFile(file: File) -> [StyleViolation] {
         return file.matchPattern("try!", withSyntaxKinds: [.Keyword]).map {
             StyleViolation(ruleDescription: self.dynamicType.description,
-                severity: config.severity,
+                severity: configuration.severity,
                 location: Location(file: file, characterOffset: $0.location))
         }
     }