+++ Source/SwiftLintFramework/Rules/StatementPositionRule.swift
@@ -9,9 +9,9 @@
 import Foundation
 import SourceKittenFramework
 
-public struct StatementPositionRule: CorrectableRule, ConfigProviderRule {
+public struct StatementPositionRule: CorrectableRule, ConfigurationProviderRule {
 
-    public var config = SeverityConfig(.Warning)
+    public var configuration = SeverityConfiguration(.Warning)
 
     public init() {}
 
@@ -44,7 +44,7 @@ public struct StatementPositionRule: CorrectableRule, ConfigProviderRule {
     public func validateFile(file: File) -> [StyleViolation] {
         return violationRangesInFile(file, withPattern: pattern).flatMap { range in
             return StyleViolation(ruleDescription: self.dynamicType.description,
-                severity: config.severity,
+                severity: configuration.severity,
                 location: Location(file: file, characterOffset: range.location))
         }
     }