+++ Source/SwiftLintFramework/Rules/TodoRule.swift
@@ -15,9 +15,9 @@ extension SyntaxKind {
     }
 }
 
-public struct TodoRule: ConfigProviderRule {
+public struct TodoRule: ConfigurationProviderRule {
 
-    public var config = SeverityConfig(.Warning)
+    public var configuration = SeverityConfiguration(.Warning)
 
     public init() {}
 
@@ -47,7 +47,7 @@ public struct TodoRule: ConfigProviderRule {
                 return nil
             }
             return StyleViolation(ruleDescription: self.dynamicType.description,
-                severity: config.severity,
+                severity: configuration.severity,
                 location: Location(file: file, characterOffset: range.location))
         }
     }