+++ Source/SwiftLintFramework/Rules/TrailingSemicolonRule.swift
@@ -16,9 +16,9 @@ extension File {
     }
 }
 
-public struct TrailingSemicolonRule: CorrectableRule, ConfigProviderRule {
+public struct TrailingSemicolonRule: CorrectableRule, ConfigurationProviderRule {
 
-    public var config = SeverityConfig(.Warning)
+    public var configuration = SeverityConfiguration(.Warning)
 
     public init() {}
 
@@ -46,7 +46,7 @@ public struct TrailingSemicolonRule: CorrectableRule, ConfigProviderRule {
     public func validateFile(file: File) -> [StyleViolation] {
         return file.violatingTrailingSemicolonRanges().map {
             StyleViolation(ruleDescription: self.dynamicType.description,
-                severity: config.severity,
+                severity: configuration.severity,
                 location: Location(file: file, characterOffset: $0.location))
         }
     }