+++ Source/SwiftLintFramework/Rules/ValidDocsRule.swift
@@ -123,9 +123,9 @@ func superfluousOrMissingParameterDocumentation(declaration: String,
     }.isEmpty
 }
 
-public struct ValidDocsRule: ConfigProviderRule {
+public struct ValidDocsRule: ConfigurationProviderRule {
 
-    public var config = SeverityConfig(.Warning)
+    public var configuration = SeverityConfiguration(.Warning)
 
     public init() {}
 
@@ -219,7 +219,7 @@ public struct ValidDocsRule: ConfigProviderRule {
     public func validateFile(file: File) -> [StyleViolation] {
         return file.invalidDocOffsets(file.structure.dictionary).map {
             StyleViolation(ruleDescription: self.dynamicType.description,
-                severity: config.severity,
+                severity: configuration.severity,
                 location: Location(file: file, byteOffset: $0))
         }
     }