+++ Source/SwiftLintFrameworkTests/ConfigurationTests.swift
@@ -42,7 +42,6 @@ class ConfigurationTests: XCTestCase {
         ("testLevel1", self.testLevel1),
         ("testLevel2", self.testLevel2),
         ("testLevel3", self.testLevel3),
-        ("testDoNotUseNestedConfigs", self.testDoNotUseNestedConfigs),
         ("testConfiguresCorrectlyFromDict", self.testConfiguresCorrectlyFromDict),
         ("testConfigureFallsBackCorrectly", self.testConfigureFallsBackCorrectly),
     ]
@@ -88,12 +87,12 @@ class ConfigurationTests: XCTestCase {
         let combinedRulesConfigDict = enabledRulesConfigDict.reduce(disabledRulesConfigDict) {
             var d = $0; d[$1.0] = $1.1; return d
         }
-        var config = Configuration(dict: enabledRulesConfigDict)
-        XCTAssertNil(config)
-        config = Configuration(dict: disabledRulesConfigDict)
-        XCTAssertNil(config)
-        config = Configuration(dict: combinedRulesConfigDict)
-        XCTAssertNil(config)
+        var configuration = Configuration(dict: enabledRulesConfigDict)
+        XCTAssertNil(configuration)
+        configuration = Configuration(dict: disabledRulesConfigDict)
+        XCTAssertNil(configuration)
+        configuration = Configuration(dict: combinedRulesConfigDict)
+        XCTAssertNil(configuration)
     }
 
     func testDisabledRules() {
@@ -154,9 +153,9 @@ class ConfigurationTests: XCTestCase {
     // MARK: - Testing Configuration Equality
 
     private var projectMockConfig0: Configuration {
-        var config = Configuration(path: projectMockYAML0, optional: false, quiet: true)
-        config.rootPath = projectMockPathLevel0
-        return config
+        var configuration = Configuration(path: projectMockYAML0, optional: false, quiet: true)
+        configuration.rootPath = projectMockPathLevel0
+        return configuration
     }
 
     private var projectMockConfig2: Configuration {
@@ -178,30 +177,23 @@ class ConfigurationTests: XCTestCase {
     }
 
     func testLevel0() {
-        XCTAssertEqual(projectMockConfig0.configForFile(File(path: projectMockSwift0)!),
+        XCTAssertEqual(projectMockConfig0.configurationForFile(File(path: projectMockSwift0)!),
                        projectMockConfig0)
     }
 
     func testLevel1() {
-        XCTAssertEqual(projectMockConfig0.configForFile(File(path: projectMockSwift1)!),
+        XCTAssertEqual(projectMockConfig0.configurationForFile(File(path: projectMockSwift1)!),
                        projectMockConfig0)
     }
 
     func testLevel2() {
-        XCTAssertEqual(projectMockConfig0.configForFile(File(path: projectMockSwift2)!),
+        XCTAssertEqual(projectMockConfig0.configurationForFile(File(path: projectMockSwift2)!),
                        projectMockConfig0.merge(projectMockConfig2))
     }
 
     func testLevel3() {
-        XCTAssertEqual(projectMockConfig0.configForFile(File(path: projectMockSwift3)!),
+        XCTAssertEqual(projectMockConfig0.configurationForFile(File(path: projectMockSwift3)!),
                        projectMockConfig0.merge(projectMockConfig2))
-    }
-
-    func testDoNotUseNestedConfigs() {
-        var config = Configuration(dict: ["use_nested_configs": false])!
-        config.rootPath = projectMockPathLevel0
-        XCTAssertEqual(config.configForFile(File(path: projectMockSwift3)!),
-                       config)
     }
 
     // MARK: - Testing Rules from config dictionary
@@ -209,11 +201,11 @@ class ConfigurationTests: XCTestCase {
     let testRuleList = RuleList(rules: RuleWithLevelsMock.self)
 
     func testConfiguresCorrectlyFromDict() {
-        let ruleConfig = [1, 2]
-        let config = [RuleWithLevelsMock.description.identifier: ruleConfig]
+        let ruleConfiguration = [1, 2]
+        let config = [RuleWithLevelsMock.description.identifier: ruleConfiguration]
         let rules = testRuleList.configuredRulesWithDictionary(config)
         // swiftlint:disable:next force_try
-        XCTAssertTrue(rules == [try! RuleWithLevelsMock(config: ruleConfig) as Rule])
+        XCTAssertTrue(rules == [try! RuleWithLevelsMock(configuration: ruleConfiguration) as Rule])
     }
 
     func testConfigureFallsBackCorrectly() {